:gear: Déplace les fonctions d'évaluation vers un module, factorise
